Predefined List for Genres and Formats
Format and Genre fields of the Book object now is in drop-down list format in Create and Edit views. Lists are defined in the views.

A few exceptions arised during the development of this stage and the program has been edited so that these are no longer the case.

However a full error-handling logic is to be implemented next.