Error Handling Implemented
I have implemented error handling for the "Create" and "Edit" pages by converting some of the Fluent API code into Data Annotations.

A few issues that occurs at the moment is that the .Count() method always returns 0 for a Genre or Author object's Books list, which makes me think that the objects are not being added into the list at all. This also might be an issue regarding the 'virtual' behaviour of the said lists. Further investigation needed.

Also there is a minor issue in regards to consistency. Most of the error messages are displayed right below their related fields except the ISBN field. For ISBN field however a JavaScript 'Alert' is in use. For the sake of keeping the look consistent, this should ideally be fixed later on.