Automatic Deletion of Authors and Genres Added
Now when we edit a book, the application checks whether the Author and Genre of the said Book is changed. If so it checks the previous value of both Genre and Author and deletes them if they have no other books in ther Books list.

This allowing of creation and deletion of Author and Genre objects from the Books controller makes it easier to manage authors and genres while also allowing me to ensure robustness. The latter part will depend on my later implementation of disabling adding/deleting of authors from their own Views.