Upload Image Implemented
Now in my 'Create' form I have a field for uploading an image for the book that's being added. The image is saved by the name of the book's ISBN code + file extension.

Any files can be added, not just images at this points. This is to be fixed later on.

Deletion of the image if the related book has been deleted is to be implemented next.