Page Heder Created
A 'Featured Titles' section is created at the top of the page to serve as a header for the page.

Page styles such as colour schemes and fonts are not to paid attention to. As this will be planned and implicated at a later stage when the backbones of the page are fully implemented.