Sections By Genre Added
A row for each one of four genres and six books for each row are added to the homepage below the header.

Right now we have unstyled headers and an image for each one of the books under their corresponding headers.

Links are empty as I do not yet have individual book pages.

Before I move on to create those, the next step I'll take is to create a suitable footer.

Again, I will not worry too much about the styling but rather the structure of it.