Books Listed by Genre and Author now Available
Now we can list books by Genre or by Author. The whole site is functional but not pretty. After creating the individual book pages in like fashion, I will work on aesthetics, and then be done with this project.

Genres and Authors links on the navigation bar now navigates to a page that simply containes a list of all the authors/genres that have books available on the site. Each name navigates to the corresponding genre/author's books list.

Index page section-by-genre names are now links to corresponding genre lists, and there is a list of genres at the end of the page for those genres that doesn't have a corresponding section in the index page.