Individual Book Pages Created
ERROR WITH LINKS
By this commit, I also attempted to navigate each appearance of a genre name to its corresponding genre page, and each apperance of an author's name to its corresponding author page. This did not work.

Somehow I can navigate to each and every genre from /Genre/Index, but I cannot navigate to genres that has commas or '&'s in them from Books/ByAuthor and Books/ByISBN pages.

Further investigation required.

I did implement individual book pages however. Again, they are functional but not pretty. This was the last major planned commit in regards to the functionality of the site. After this (apart from fixing the error with links) I will work on page aesthetics and layout.

Finally, I will clean the code and the app will be good to publish on my portfolio.