Individual Book Pages Edited to Full Functionality
ByISBN pages now display the book image on one side and all the information about the book on the other.

More importantly, the error with the mapping to certain genre pages has been fixed. Everything in the page now works as it should.

Only the aesthetics and responsiveness of the page are yet to be implemented.