Nested For Loop for Sections
I did NOT work on the responsiveness of the page in this commit.  Rather I have worked on the code for displaying books on the homepage.

Hard coded sections are replaced by a nested for loop displaying six books for each section in the likemanner as it did before. Nested loop is fed an array of arrays. First array represents each section and second array each book.

I shall work on the responsiveness of the page on the next commit.