Found and NotFound Bottlenecked
Now both Found and NotFound views depend on the Search action of the Books controller to be called. From the pages that has the SearchBar, parameters are passed to the Search action. Then the Search action works out what view is to be called and set of books to be passed to that view;