Seven New ViewModels Added
In order to fully incorporate ViewModels into the app, seven ViewModels have been added.

There is a minor issue with the generation of the genre names that are not featured as a six column section in the HomePage. Fix that next.

You should also check whether any View is depending on a Model. If so, change this for it to depend on a ViewModel.