HomeController and HomeIndexVM are Clean
I have moved a few functions that we are using to populate a HomeIndexViewModel elswhere. Two of then to the Genre model and one to the Book model, according to their relevance. Although, it seems probable to me that a better approach would be to change these for extension methods later on.